requirements.txt
sentence-transformers
numpy
pandas
faiss-cpu
google-generativeai
tqdm
scikit-learn
psycopg2-binary
python-dotenv
